requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.4
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
databases==0.9.0
asyncpg==0.30.0fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.4
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
databases==0.9.0
asyncpg==0.30.0